requirements.txt
# Agentic AI Retail System - Dependencies

# Core
python-dotenv>=1.0.0
pydantic>=2.0.0

# Groq LLM
groq>=0.4.0

# Supabase Database
supabase>=2.0.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# HTTP Client
httpx>=0.25.0

# Utilities
python-dateutil>=2.8.0

# =============================================================================
# VOICE (Deepgram - STT/TTS)
# =============================================================================
# Deepgram uses httpx for REST API, already included above
# No additional dependencies needed for basic Deepgram integration
